sistema_financeiro v0.1.0 Money

Represents a Money type, inspired in Martin Fowler's Money Patter

Functions

add(a, b)

Raises different currencies exception

Examples:

  iex> Money.add(Money.new(10), Money.new(10, :USD))
  ** (ArgumentError) Monies with different currencies. Got BRL and USD

divide(m, ratios)

Divides Money from a given a denominator

Examples:

  iex> Money.divide(Money.new(10), 2)
  [%Money{amount: 500, currency: :BRL}, %Money{amount: 500, currency: :BRL}]
  iex> Money.divide(Money.new(9), 3)
  [%Money{amount: 300, currency: :BRL}, %Money{amount: 300, currency: :BRL}, %Money{amount: 300, currency: :BRL}]
  iex> Money.divide(Money.new(9, :USD), 1)
  [%Money{amount: 900, currency: :USD}]
  iex> Money.divide(Money.new(5), "2")
  ** (ArgumentError) Value "2" must be integer

multiply(a, b)

Multiplies Money by amount

Examples:

  iex> Money.multiply(Money.new(15, :USD), 2)
  %Money{amount: 3000, currency: :USD}
  iex> Money.multiply(Money.new(750, :JPY), 3.5)
  %Money{amount: 2625, currency: :JPY}
  iex> Money.multiply(Money.new(750), "3.5")
  ** (ArgumentError) Value "3.5" must be integer or float

multiply(a, b, currency)

Multiplies Money by amount and create a new one with a given Currency

Examples:

  iex> Money.multiply(Money.new(15), 2, :USD)
  %Money{amount: 3000, currency: :USD}
  iex> Money.multiply(Money.new(750, :USD), 3.5, :JPY)
  %Money{amount: 2625, currency: :JPY}

new(amount, currency_code \\ :BRL)

Creates a new Money type with amount and currency: Default currency is :BRL

Examples:

  iex> Money.new(5)
  %Money{amount: 500, currency: :BRL}
  iex> Money.new(5, :USD)
  %Money{amount: 500, currency: :USD}
  iex> Money.new(5.78, :USD)
  %Money{amount: 578, currency: :USD}
  iex> Money.new(5, "USD")
  %Money{amount: 500, currency: :USD}
  iex> Money.new(5, "usd")
  %Money{amount: 500, currency: :USD}
  iex> Money.new(5, "new_currency")
  {:error, "Currency NEW_CURRENCY not found"}

new!(amount, currency_code \\ :BRL)

Creates a new Money type with amount and currency: Default currency is :BRL

Examples:

  iex> Money.new!(5)
  %Money{amount: 500, currency: :BRL}
  iex> Money.new!(5, :USD)
  %Money{amount: 500, currency: :USD}
  iex> Money.new!(5.78, :USD)
  iex> Money.new!(5, "new_currency")
  ** (ArgumentError) Currency NEW_CURRENCY not found

parse(amount, currency \\ :BRL)

Parse an amount to Money

Examples:

  iex> Money.parse("12")
  %Money{amount: 1200, currency: :BRL}
  iex> Money.parse("0,1")
  %Money{amount: 10, currency: :BRL}
  iex> Money.parse("12aa", :USD)
  {:error, "Cannot parse value \"12aa\""}

parse!(amount, currency \\ :BRL)

Parse an amount value to Money. Raises an error if the value is not a number

Examples:

  iex> Money.parse!("12")
  %Money{amount: 1200, currency: :BRL}
  iex> Money.parse!("0.1", :USD)
  %Money{amount: 10, currency: :USD}
  iex> Money.parse!("0,1")
  %Money{amount: 10, currency: :BRL}
  iex> Money.parse!("bad", :USD)
  ** (ArgumentError) Cannot parse value "bad"

to_string(m)

Converts Money to formated string with properly symbol and number of decimal cases

Examples:

  iex> Money.to_string(Money.new(4))
  "R$ 4.00"
  iex> Money.to_string(Money.new(25, :USD))
  "$ 25.00"