sistema_financeiro v0.1.0 SistemaFinanceiro

Core module of the app where you can start the application and play around with the Accounts

Functions

start()

Starts the app by listing all accounts

Examples

  iex> SistemaFinanceiro.start()
  ################################
              Accounts
  ################################
  Code     Owner          Balance
  1        Zé Doe         R$ 50.00

  2        Zeller Doe     R$ 50.00

  3        Ann Doe        $ 25.00

  4        Smith Doe      $ 100.00

  5        Xang Doe       $ 25.00

  6        Zang Doe       ¥ 50

  7        Xan Jack       ¥ 25

  (L)ist Accounts (S)plit Money (E)xchange (Q)uit: